+++ src/index.js
@@ -1,28 +1,36 @@
-import { indexOf } from 'lodash';
+import { taskNotComplete, taskCompleted } from './status.js';
 import './style.css';
-// const tasks = [
-//   {
-//     description: 'Complete list structure',
-//     completed: false,
-//     index: 1,
-//   },
-//   {
-//     description: 'Complete interactive list task',
-//     completed: false,
-//     index: 2,
-//   },
-//   {
-//     description: 'Work on adding and removing tasks',
-//     completed: false,
-//     index: 3,
-//   },
-// ];
 
- 
- const tasks = getFromLocalStorage();
+// eslint-disable-no-use-before-define import/prefer-default-export
+const tasks1 = [
+  {
+    description: 'Complete list structure',
+    completed: false,
+    index: 1,
+  },
+  {
+    description: 'Complete interactive list task',
+    completed: false,
+    index: 2,
+  },
+  {
+    description: 'Work on adding and removing tasks',
+    completed: false,
+    index: 3,
+  },
+];
 
-const completedTasks = [];
+function getFromLocalStorage() {
+  const storage = JSON.parse(localStorage.getItem('lstore'));
+  return storage;
+}
 
+const tasks = getFromLocalStorage() || tasks1;
+
+function addToLocalStorage() {
+  const storage = JSON.stringify(tasks);
+  localStorage.setItem('lstore', storage);
+}
 
 const ul = document.querySelector('ul');
 function ui() {
@@ -41,8 +49,6 @@ function ui() {
     checkbox.value = `${tasks[i]}`;
     checkbox.classList.add('chkbx');
 
- 
-
     li.appendChild(lDiv);
 
     lDiv.appendChild(lDivSpan);
@@ -54,73 +60,26 @@ function ui() {
 
     ul.appendChild(li);
 
-    if(tasks[i].completed == true){
+    if (tasks[i].completed === true) {
       pTask.innerText = `${tasks[i].description}`;
       pTask.classList.add('cancel');
-      console.log( pTask.previousSibling.children[0])
       pTask.previousSibling.children[0].checked = true;
-      
-    }else{
+    } else {
       pTask.innerText = `${tasks[i].description}`;
     }
 
-    
     checkbox.addEventListener('change', (e) => {
       if (e.target.checked) {
         taskCompleted(tasks[i]);
         e.target.parentElement.nextSibling.classList.add('cancel');
+        addToLocalStorage();
       } else {
         taskNotComplete(tasks[i]);
         e.target.parentElement.nextSibling.classList.remove('cancel');
-
-
+        addToLocalStorage();
       }
     });
-
   }
-
 }
-
-function taskCompleted(task){
-  task.completed = true;
-  completedTasks.push(task);
-  // console.log(completedTasks)
-  // console.log(tasks);
-  addToLocalStorage()
-
-}
-
-function taskNotComplete(task){
-  let i = indexOf(task);
-  task.completed = false;
-  completedTasks.splice(i, 1);
-  // console.log(completedTasks)
-  // console.log(tasks);
-  addToLocalStorage()
-}
-
-function addToLocalStorage(){
-const storage = JSON.stringify(tasks);
-localStorage.setItem('lstore', storage);
-}
-
-function getFromLocalStorage(){
-  const storage = JSON.parse(localStorage.getItem('lstore')) || tasks;
-  return storage;
-}
-
-// document.addEventListener('DOMContentLoaded', () => {
-//   const tasks = getFromLocalStorage();
-//   for(let i=0; i<fromStorage.length; i++){
-//     if(fromStorage[i].completed == true){
-      
-//     }
-//   }
-// })
 
 ui();
-
-
-
-
-
