+++ src/style.css
@@ -5,7 +5,7 @@
 
 .cancel {
   text-decoration: line-through;
-  color:rgb(182, 165, 165);
+  color: rgb(182, 165, 165);
 }
 
 body {
@@ -56,4 +56,3 @@ li {
   background-color: rgb(243, 241, 241);
   color: rgb(128, 115, 115);
 }
-
